refactored factories and utils in Domain.UnitTests.Users
